backend/requirements.txt
fastapi
starlette
Pillow
